Many edits.
Finished the Skeleton Cave and Damp Lair.
Added grammar to Creature so it can tell when to use "the", "a", "an", or nothing at all for unique monsters.
Added optional proc function to gear so it can do something special on a chance.
Made gear able to add to more than one stat when equipped.